Fix sijoittelun-tulos route error
